More error handling. fix bot crashing when !destroy although there is no active instance

+++ src/destroyInstance.js
@@ -1,15 +1,26 @@
 const vultr = require("./vultrinitializer");
 const Discord = require("discord.js");
+const { vultrAPI } = require("../config.json");
 
 function destroyInstance(message) {
   let InstanceID;
+  if (!vultrAPI) {
+    message.channel.send(
+      `There is no instance to destroy since your API token is not defined!`
+    );
+    return;
+  }
   vultr.api.instances.listInstances().then((res) => {
-    InstanceID = res.instances[0].id;
-
-    vultr.api.instances.deleteInstance({
-      "instance-id": `${InstanceID}`,
-    });
-
+    console.log(res.instances.length);
+    if (res.instances.length === 0) {
+      message.channel.send(`There are not active instances currently`);
+      return;
+    } else {
+      InstanceID = res.instances[0].id;
+      vultr.api.instances.deleteInstance({
+        "instance-id": `${InstanceID}`,
+      });
+    }
     let embedMessageDestroyed = new Discord.MessageEmbed()
       .setTitle("Server Destroyed!")
       .setDescription(`Server with id: ${InstanceID} is destroyed!`)
